Made it possible to build again for windows
